feat: utm generator pages for individual channels
